renamed CrudFeatures folder to Operations and combined files so that there's one file per operation
