+++ src/app/Http/Controllers/CrudController.php
@@ -10,16 +10,18 @@ use Illuminate\Foundation\Bus\DispatchesJobs;
 use Illuminate\Http\Request as UpdateRequest;
 use Illuminate\Routing\Controller as BaseController;
 use Illuminate\Foundation\Validation\ValidatesRequests;
-use Backpack\CRUD\app\Http\Controllers\CrudFeatures\Reorder;
-use Backpack\CRUD\app\Http\Controllers\CrudFeatures\AjaxTable;
-use Backpack\CRUD\app\Http\Controllers\CrudFeatures\Revisions;
-use Backpack\CRUD\app\Http\Controllers\CrudFeatures\SaveActions;
-use Backpack\CRUD\app\Http\Controllers\CrudFeatures\ShowDetailsRow;
+use Backpack\CRUD\app\Http\Controllers\Operations\Create;
+use Backpack\CRUD\app\Http\Controllers\Operations\ListEntries;
+use Backpack\CRUD\app\Http\Controllers\Operations\Reorder;
+use Backpack\CRUD\app\Http\Controllers\Operations\Revisions;
+use Backpack\CRUD\app\Http\Controllers\Operations\SaveActions;
+use Backpack\CRUD\app\Http\Controllers\Operations\Show;
+use Backpack\CRUD\app\Http\Controllers\Operations\Update;
 
 class CrudController extends BaseController
 {
     use DispatchesJobs, ValidatesRequests;
-    use AjaxTable, Reorder, Revisions, ShowDetailsRow, SaveActions;
+    use Create, Update, ListEntries, Show, Reorder, Revisions, SaveActions;
 
     public $data = [];
 
@@ -54,183 +56,6 @@ class CrudController extends BaseController
     {
     }
 
-    /**
-     * Display all rows in the database for this entity.
-     *
-     * @return Response
-     */
-    public function index()
-    {
-        $this->crud->hasAccessOrFail('list');
-
-        $this->data['crud'] = $this->crud;
-        $this->data['title'] = ucfirst($this->crud->entity_name_plural);
-
-        // load the view from /resources/views/vendor/backpack/crud/ if it exists, otherwise load the one in the package
-        return view($this->crud->getListView(), $this->data);
-    }
-
-    /**
-     * Show the form for creating inserting a new row.
-     *
-     * @return Response
-     */
-    public function create()
-    {
-        $this->crud->hasAccessOrFail('create');
-
-        // prepare the fields you need to show
-        $this->data['crud'] = $this->crud;
-        $this->data['saveAction'] = $this->getSaveAction();
-        $this->data['fields'] = $this->crud->getCreateFields();
-        $this->data['title'] = trans('backpack::crud.add').' '.$this->crud->entity_name;
-
-        // load the view from /resources/views/vendor/backpack/crud/ if it exists, otherwise load the one in the package
-        return view($this->crud->getCreateView(), $this->data);
-    }
-
-    /**
-     * Store a newly created resource in the database.
-     *
-     * @param StoreRequest $request - type injection used for validation using Requests
-     *
-     * @return \Illuminate\Http\RedirectResponse
-     */
-    public function storeCrud(StoreRequest $request = null)
-    {
-        $this->crud->hasAccessOrFail('create');
-
-        // fallback to global request instance
-        if (is_null($request)) {
-            $request = \Request::instance();
-        }
-
-        // replace empty values with NULL, so that it will work with MySQL strict mode on
-        foreach ($request->input() as $key => $value) {
-            if (empty($value) && $value !== '0') {
-                $request->request->set($key, null);
-            }
-        }
-
-        // insert item in the db
-        $item = $this->crud->create($request->except(['save_action', '_token', '_method', 'current_tab']));
-        $this->data['entry'] = $this->crud->entry = $item;
-
-        // show a success message
-        \Alert::success(trans('backpack::crud.insert_success'))->flash();
-
-        // save the redirect choice for next time
-        $this->setSaveAction();
-
-        return $this->performSaveAction($item->getKey());
-    }
-
-    /**
-     * Show the form for editing the specified resource.
-     *
-     * @param int $id
-     *
-     * @return Response
-     */
-    public function edit($id)
-    {
-        $this->crud->hasAccessOrFail('update');
-
-        // get entry ID from Request (makes sure its the last ID for nested resources)
-        $id = $this->crud->getCurrentEntryId() ?? $id;
-
-        // get the info for that entry
-        $this->data['entry'] = $this->crud->getEntry($id);
-        $this->data['crud'] = $this->crud;
-        $this->data['saveAction'] = $this->getSaveAction();
-        $this->data['fields'] = $this->crud->getUpdateFields($id);
-        $this->data['title'] = trans('backpack::crud.edit').' '.$this->crud->entity_name;
-
-        $this->data['id'] = $id;
-
-        // load the view from /resources/views/vendor/backpack/crud/ if it exists, otherwise load the one in the package
-        return view($this->crud->getEditView(), $this->data);
-    }
-
-    /**
-     * Update the specified resource in the database.
-     *
-     * @param UpdateRequest $request - type injection used for validation using Requests
-     *
-     * @return \Illuminate\Http\RedirectResponse
-     */
-    public function updateCrud(UpdateRequest $request = null)
-    {
-        $this->crud->hasAccessOrFail('update');
-
-        // fallback to global request instance
-        if (is_null($request)) {
-            $request = \Request::instance();
-        }
-
-        // replace empty values with NULL, so that it will work with MySQL strict mode on
-        foreach ($request->input() as $key => $value) {
-            if (empty($value) && $value !== '0') {
-                $request->request->set($key, null);
-            }
-        }
-
-        // update the row in the db
-        $item = $this->crud->update($request->get($this->crud->model->getKeyName()),
-                            $request->except('save_action', '_token', '_method', 'current_tab'));
-        $this->data['entry'] = $this->crud->entry = $item;
-
-        // show a success message
-        \Alert::success(trans('backpack::crud.update_success'))->flash();
-
-        // save the redirect choice for next time
-        $this->setSaveAction();
-
-        return $this->performSaveAction($item->getKey());
-    }
-
-    /**
-     * Display the specified resource.
-     *
-     * @param int $id
-     *
-     * @return Response
-     */
-    public function show($id)
-    {
-        $this->crud->hasAccessOrFail('show');
-
-        // get entry ID from Request (makes sure its the last ID for nested resources)
-        $id = $this->crud->getCurrentEntryId() ?? $id;
-
-        // set columns from db
-        $this->crud->setFromDb();
-
-        // cycle through columns
-        foreach ($this->crud->columns as $key => $column) {
-            // remove any autoset relationship columns
-            if (array_key_exists('model', $column) && array_key_exists('autoset', $column) && $column['autoset']) {
-                $this->crud->removeColumn($column['name']);
-            }
-
-            // remove the row_number column, since it doesn't make sense in this context
-            if ($column['type'] == 'row_number') {
-                $this->crud->removeColumn($column['name']);
-            }
-        }
-
-        // get the info for that entry
-        $this->data['entry'] = $this->crud->getEntry($id);
-        $this->data['crud'] = $this->crud;
-        $this->data['title'] = trans('backpack::crud.preview').' '.$this->crud->entity_name;
-
-        // remove preview button from stack:line
-        $this->crud->removeButton('preview');
-        $this->crud->removeButton('delete');
-
-        // load the view from /resources/views/vendor/backpack/crud/ if it exists, otherwise load the one in the package
-        return view($this->crud->getShowView(), $this->data);
-    }
 
     /**
      * Remove the specified resource from storage.
