--- src/app/Http/Controllers/CrudFeatures/ShowDetailsRow.php
@@ -1,29 +1,0 @@
-<?php
-
-namespace Backpack\CRUD\app\Http\Controllers\CrudFeatures;
-
-trait ShowDetailsRow
-{
-    /**
-     * Used with AJAX in the list view (datatables) to show extra information about that row that didn't fit in the table.
-     * It defaults to showing some dummy text.
-     *
-     * It's enabled by:
-     * - setting: $crud->details_row = true;
-     * - adding the details route for the entity; ex: Route::get('page/{id}/details', 'PageCrudController@showDetailsRow');
-     * - adding a view with the following name to change what the row actually contains: app/resources/views/vendor/backpack/crud/details_row.blade.php
-     */
-    public function showDetailsRow($id)
-    {
-        $this->crud->hasAccessOrFail('details_row');
-
-        // get entry ID from Request (makes sure its the last ID for nested resources)
-        $id = $this->crud->getCurrentEntryId() ?? $id;
-
-        $this->data['entry'] = $this->crud->getEntry($id);
-        $this->data['crud'] = $this->crud;
-
-        // load the view from /resources/views/vendor/backpack/crud/ if it exists, otherwise load the one in the package
-        return view($this->crud->getDetailsRowView(), $this->data);
-    }
-}
