+++ src/app/Http/Controllers/Operations/ListEntries.php
@@ -1,9 +1,25 @@
 <?php
 
-namespace Backpack\CRUD\app\Http\Controllers\CrudFeatures;
+namespace Backpack\CRUD\app\Http\Controllers\Operations;
 
-trait AjaxTable
+trait ListEntries
 {
+    /**
+     * Display all rows in the database for this entity.
+     *
+     * @return Response
+     */
+    public function index()
+    {
+        $this->crud->hasAccessOrFail('list');
+
+        $this->data['crud'] = $this->crud;
+        $this->data['title'] = ucfirst($this->crud->entity_name_plural);
+
+        // load the view from /resources/views/vendor/backpack/crud/ if it exists, otherwise load the one in the package
+        return view($this->crud->getListView(), $this->data);
+    }
+
     /**
      * The search function that is called by the data table.
      *
@@ -15,7 +31,6 @@ trait AjaxTable
 
         $totalRows = $filteredRows = $this->crud->count();
         $startIndex = $this->request->input('start') ?: 0;
-
         // if a search term was present
         if ($this->request->input('search') && $this->request->input('search')['value']) {
             // filter the results accordingly
@@ -23,23 +38,19 @@ trait AjaxTable
             // recalculate the number of filtered rows
             $filteredRows = $this->crud->count();
         }
-
         // start the results according to the datatables pagination
         if ($this->request->input('start')) {
             $this->crud->skip($this->request->input('start'));
         }
-
         // limit the number of results according to the datatables pagination
         if ($this->request->input('length')) {
             $this->crud->take($this->request->input('length'));
         }
-
         // overwrite any order set in the setup() method with the datatables order
         if ($this->request->input('order')) {
             $column_number = $this->request->input('order')[0]['column'];
             $column_direction = $this->request->input('order')[0]['dir'];
             $column = $this->crud->findColumnById($column_number);
-
             if ($column['tableColumn']) {
                 // clear any past orderBy rules
                 $this->crud->query->getQuery()->orders = null;
@@ -47,9 +58,30 @@ trait AjaxTable
                 $this->crud->orderBy($column['name'], $column_direction);
             }
         }
-
         $entries = $this->crud->getEntries();
-
         return $this->crud->getEntriesAsJsonForDatatables($entries, $totalRows, $filteredRows, $startIndex);
+    }
+
+    /**
+     * Used with AJAX in the list view (datatables) to show extra information about that row that didn't fit in the table.
+     * It defaults to showing some dummy text.
+     *
+     * It's enabled by:
+     * - setting: $crud->details_row = true;
+     * - adding the details route for the entity; ex: Route::get('page/{id}/details', 'PageCrudController@showDetailsRow');
+     * - adding a view with the following name to change what the row actually contains: app/resources/views/vendor/backpack/crud/details_row.blade.php
+     */
+    public function showDetailsRow($id)
+    {
+        $this->crud->hasAccessOrFail('details_row');
+
+        // get entry ID from Request (makes sure its the last ID for nested resources)
+        $id = $this->crud->getCurrentEntryId() ?? $id;
+
+        $this->data['entry'] = $this->crud->getEntry($id);
+        $this->data['crud'] = $this->crud;
+
+        // load the view from /resources/views/vendor/backpack/crud/ if it exists, otherwise load the one in the package
+        return view($this->crud->getDetailsRowView(), $this->data);
     }
 }
