+++ src/app/Http/Controllers/Operations/Show.php
@@ -1,0 +1,50 @@
+<?php
+
+namespace Backpack\CRUD\app\Http\Controllers\Operations;
+
+trait Show
+{
+    /**
+     * Display the specified resource.
+     *
+     * @param int $id
+     *
+     * @return Response
+     */
+    public function show($id)
+    {
+        $this->crud->hasAccessOrFail('show');
+
+        // get entry ID from Request (makes sure its the last ID for nested resources)
+        $id = $this->crud->getCurrentEntryId() ?? $id;
+
+        // set columns from db
+        $this->crud->setFromDb();
+
+        // cycle through columns
+        foreach ($this->crud->columns as $key => $column) {
+            // remove any autoset relationship columns
+            if (array_key_exists('model', $column) && array_key_exists('autoset', $column) && $column['autoset']) {
+                $this->crud->removeColumn($column['name']);
+            }
+
+            // remove the row_number column, since it doesn't make sense in this context
+            if ($column['type'] == 'row_number') {
+                $this->crud->removeColumn($column['name']);
+            }
+        }
+
+        // get the info for that entry
+        $this->data['entry'] = $this->crud->getEntry($id);
+        $this->data['crud'] = $this->crud;
+        $this->data['title'] = trans('backpack::crud.preview').' '.$this->crud->entity_name;
+
+        // remove preview button from stack:line
+        $this->crud->removeButton('preview');
+        $this->crud->removeButton('delete');
+
+        // load the view from /resources/views/vendor/backpack/crud/ if it exists, otherwise load the one in the package
+        return view($this->crud->getShowView(), $this->data);
+    }
+
+}
